% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurt.R
\name{kurt}
\alias{M4}
\alias{PortKurt}
\alias{PortKurtContrib}
\alias{PortKurtDeriv}
\alias{cm4}
\alias{dm4}
\alias{kurt}
\alias{pm4}
\title{Fourth centered moments}
\usage{
M4(r)

pm4(r, w)

dm4(r, w)

cm4(r, w, percentage = TRUE)

PortKurt(r, w)

PortKurtDeriv(r, w)

PortKurtContrib(r, w, percentage = TRUE)
}
\arguments{
\item{r}{\code{matrix}, a \eqn{(T x N)} array of returns.}

\item{w}{\code{numeric}, a \eqn{(N x 1)} vector of portfolio weights.}

\item{percentage}{\code{logical}, whether risk contributions are expressed as percentages.}
}
\value{
\code{numeric}
}
\description{
These functions relate to the computation of the kurtosis (\code{M4()}),
the portfolio variance risk (\code{pm4()} and \code{PortKurt()}),
the partial derivatives (\code{dm4()} and \code{PortKurtDeriv()}) and
the risk contributions of the assets (\code{cm4()} and \code{PortKurtContrib()}).
}
\examples{
data(MultiAsset)
MA <- as.timeSeries(MultiAsset[, 1:4])
r <- na.omit(diff(log(MA)) * 100)
N <- ncol(r)
w <- rep(1 / N, N) ## equal weight allocation
M4(r)
pm4(r, w)
dm4(r, w)
}
\references{
Boudt, K. and Peterson, B. and Croux, C. (2008/09),
Estimation and decomposition of downside risk for portfolios with non-normal returns,
\emph{The Journal of Risk}, \bold{11}(2), Winter 2008/09, 79--103.

Jondeau, E. and Rockinger, M. (2006), Optimal portfolio allocation under higher moments,
\emph{European Financial Management}, \bold{12}(1), 29--55.
}
\seealso{
Other res: \code{\link{risk}}, \code{\link{skew}}
}

