% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcrp.R
\name{mcrp}
\alias{mcrp}
\title{Multiple criteria risk parity optimization}
\usage{
mcrp(start, returns, lambda = c(1, 1, 1), ...)
}
\arguments{
\item{start}{\code{numeric}, a \eqn{(N x 1)} vector of starting values.}

\item{returns}{\code{matrix}, a \eqn{(T x N)} matrix of asset returns.}

\item{lambda}{\code{numeric}, a \eqn{(3 x 1)} weight vector for the partial objectives.}

\item{...}{ellipsis argument, passed to \code{nlminb()}.}
}
\value{
\code{list}
}
\description{
This function conducts a multiple criteria optimization with respect to the assets'
risk contributions of the portfolio higher moments (variance, skewness and kurtosis).
The three-element weight vector \code{lambda} determines with respect to which higher moments the
optimization is conducted. \code{NA} entry(ies) in this vector indicate the skipping of
a higher moment. That is \code{lambda = c(1, NA, NA)} would conduct a pure ERC optimization
with respect to the assets' contribution to the portfolio variance.
}
\examples{
data(EuStockMarkets)
r <- diff(log(EuStockMarkets)) * 100
N <- ncol(r)
w <- rep(1 / N, N) ## start values
erc <- mcrp(start = w, returns = r, lambda = c(1, NA, NA), lower = rep(0, N))
w <- Weights(erc)
PortRiskContrib(r, w)

}
\references{
Baitinger, E. and Dragosch, A. and Topalova, A. (2017),
Extending the Risk Parity Approach to Higher Moments: Is there Any Value Added?,
\emph{The Journal of Portfolio Managament}, \bold{43}(2), 24--36.

Boudt, K. and Peterson, B. and Croux, C. (2008/09), Estimation and decomposition of downside risk
for portfolios with non-normal returns, \emph{The Journal of Risk}, \bold{11}(2), Winter 2008/09, 79--103.

Jondeau, E. and Rockinger, M. (2006), Optimal portfolio allocation under higher moments, \
emph{European Financial Management}, \bold{12}(1), 29--55.
}

