% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk.R
\name{risk}
\alias{M2}
\alias{PortRisk}
\alias{PortRiskContrib}
\alias{PortRiskDeriv}
\alias{cm2}
\alias{dm2}
\alias{pm2}
\alias{risk}
\title{Second centered moments}
\usage{
M2(r)

pm2(r, w)

dm2(r, w)

cm2(r, w, percentage = TRUE)

PortRisk(r, w)

PortRiskDeriv(r, w)

PortRiskContrib(r, w, percentage = TRUE)
}
\arguments{
\item{r}{\code{matrix}, a \eqn{(T x N)} array of returns.}

\item{w}{\code{numeric}, a \eqn{(N x 1)} vector of portfolio weights.}

\item{percentage}{\code{logical}, whether risk contributions are expressed as percentages.}
}
\value{
\code{numeric}
}
\description{
These functions relate to the computation of second centered moments (\code{M2()}),
the portfolio variance risk (\code{pm2()} and \code{PortRisk()}),
the partial derivatives (\code{dm2()} and \code{PortRiskDeriv()}) and
the risk contributions of the assets (\code{cm2()} and \code{PortRiskContrib()}).
}
\examples{
data(MultiAsset)
MA <- as.timeSeries(MultiAsset[, 1:4])
r <- na.omit(diff(log(MA)) * 100)
N <- ncol(r)
w <- rep(1 / N, N) ## equal weight allocation
M2(r)
pm2(r, w)
dm2(r, w)
}
\references{
Boudt, K. and Peterson, B. and Croux, C. (2008/09), Estimation and decomposition
of downside risk for portfolios with non-normal returns,
\emph{The Journal of Risk}, \bold{11}(2), Winter 2008/09, 79--103.

Jondeau, E. and Rockinger, M. (2006), Optimal portfolio allocation under higher moments,
\emph{European Financial Management}, \bold{12}(1), 29--55.
}
\seealso{
Other res: \code{\link{kurt}}, \code{\link{skew}}
}

