% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{M3}
\alias{PortSkew}
\alias{PortSkewContrib}
\alias{PortSkewDeriv}
\alias{cm3}
\alias{dm3}
\alias{pm3}
\alias{skew}
\title{Third centered moments}
\usage{
M3(r)

pm3(r, w)

dm3(r, w)

cm3(r, w, percentage = TRUE)

PortSkew(r, w)

PortSkewDeriv(r, w)

PortSkewContrib(r, w, percentage = TRUE)
}
\arguments{
\item{r}{\code{matrix}, a \eqn{(T x N)} array of returns.}

\item{w}{\code{numeric}, a \eqn{(N x 1)} vector of portfolio weights.}

\item{percentage}{\code{logical}, whether risk contributions are expressed as percentages.}
}
\value{
\code{numeric}
}
\description{
These functions relate to the computation of (co-)skewness (\code{M3()}),
the portfolio skewness (\code{pm3()} and \code{PortSkew()}), the partial
derivatives (\code{dm3()} and \code{PortSkewDeriv()})
and the contributions of the assets to skewness (\code{cm3()} and
\code{PortSkewContrib()}).
}
\examples{
data(EuStockMarkets)
r <- diff(log(EuStockMarkets)) * 100
N <- ncol(r)
w <- rep(1 / N, N) ## equal weight allocation
M3(r)
pm3(r, w)
dm3(r, w)
}
\references{
Boudt, K. and Peterson, B. and Croux, C. (2008/09), Estimation and decomposition of
downside risk for portfolios with non-normal returns, \emph{The Journal of Risk}, \bold{11}(2),
Winter 2008/09, 79--103.

Jondeau, E. and Rockinger, M. (2006), Optimal portfolio allocation under higher moments,
\emph{European Financial Management}, \bold{12}(1), 29--55.
}
\seealso{
Other res: \code{\link{kurt}}, \code{\link{risk}}
}

